Add files via upload
changelog 0.19.0
- implemented signs (create a sign in the first line with [ix] then click with a item on it)
- moved from float to doulbe at prices for better precision
- added
    currencySymbol
    decimals
    decimal_separator
    thousand_separator
    unitLocation
    to config.yml for formatted price

- remove some debugging messages

+++ Itemex/src/main/java/sh/ome/itemex/RAM/TopOrders.java
@@ -3,25 +3,25 @@ package sh.ome.itemex.RAM;
 // consumes at 1500 entries = ( 4*8 + 4*8 ) * 1500 = 432,000byte = 421.875kbyte
 
 public class TopOrders {
-    private float[] top_buyorder_prices = new float[5];
-    private float[] top_sellorder_prices = new float[5];
+    private double[] top_buyorder_prices = new double[5];
+    private double[] top_sellorder_prices = new double[5];
     private int[] buyorder_amount = new int[5];
     private int[] sellorder_amount = new int[5];
 
     // Constructor
     public TopOrders () {}
-    public TopOrders(float[] top_buyorder_prices, float[] top_sellorder_prices, int[] buyorder_amount, int[] sellorder_amount) {
+    public TopOrders(double[] top_buyorder_prices, double[] top_sellorder_prices, int[] buyorder_amount, int[] sellorder_amount) {
         this.top_buyorder_prices = top_buyorder_prices;
         this.top_sellorder_prices = top_sellorder_prices;
         this.buyorder_amount = buyorder_amount;
         this.sellorder_amount = sellorder_amount;
     }
 
-    public float[] get_top_buyorder_prices() {
+    public double[] get_top_buyorder_prices() {
         return this.top_buyorder_prices;
     }
 
-    public float[] get_top_sellorder_prices() {
+    public double[] get_top_sellorder_prices() {
         return this.top_sellorder_prices;
     }
